requirements.txt
discord.py
uszipcode